+++ changelog.md
@@ -1,2 +1,17 @@
-# v1.0.0
+# <cj>v1.0.1</c>
+- Added stats to Official Levels:
+	- Main
+	- Tower
+	- Meltdown
+	- World
+	- SubZero
+	- Other
+- Added stat to User Coins:
+	- Unverified Coins
+- Added stat to Collected Diamonds:
+	- Diamond Shards
+- Added stat to Collected Orbs:
+	- Current Orbs
+
+# <cl>v1.0.0</c>
 - Initial release of Improved Stats